refactor(ui): split StepInspector and move it onto colour tokens
795 -> 526 lines; the decision narrative and signal rows move to
StepInspectorSignals.tsx. Also fixes an a11y warning on the theme toggle:
aria-selected is invalid on role=button, so it now uses aria-pressed.
